Drag&Drop to graphics view address fields, navigate & delete breakpoints

+++ src/views/BreakpointView.h
@@ -3,7 +3,7 @@ struct UserData;
 
 struct BreakpointView {
 	bool open;
-	int selected_row;
+	size_t selected_row;
 
 	BreakpointView();
 	void WriteConfig(UserData& config);
